requirements.txt
scikit-learn==1.3.1
fastapi==0.68.0
python-multipart==0.0.5
uvicorn==0.15.0
gunicorn==20.1.0
pandas
numpy
Pillow
colorthief
webcolors
statistics
